+++ fhirconnect/src/main/java/com/syntaric/openfhir/fc/schema/context/ContextProfile.java
@@ -5,13 +5,15 @@ import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import lombok.Data;
 
+import java.io.Serializable;
+
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonPropertyOrder({
         "url",
         "version"
 })
 @Data
-public class ContextProfile {
+public class ContextProfile implements Serializable {
 
     @JsonProperty("url")
     private String url;