+++ fhirconnect/src/main/java/com/syntaric/openfhir/fc/schema/model/FhirConfig.java
@@ -5,6 +5,7 @@ package com.syntaric.openfhir.fc.schema.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.io.Serializable;
 
 
 /**
@@ -16,7 +17,7 @@ import com.fasterxml.jackson.annotation.JsonPropertyOrder;
         "structureDefinition"
 })
 
-public class FhirConfig {
+public class FhirConfig implements Serializable {
 
     /**
      * (Required)