+++ fhirconnect/src/main/java/com/syntaric/openfhir/fc/schema/model/SplitModel.java
@@ -3,6 +3,7 @@ package com.syntaric.openfhir.fc.schema.model;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.io.Serializable;
 import java.util.List;
 import lombok.Data;
 
@@ -13,7 +14,7 @@ import lombok.Data;
         "unique"
 })
 @Data
-public class SplitModel {
+public class SplitModel implements Serializable {
 
     @JsonProperty("create")
     private String create;