+++ fhirconnect/src/main/java/com/syntaric/openfhir/fc/schema/terminology/TerminologyMappings.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import com.syntaric.openfhir.fc.schema.terminology.TerminologyMappingsSystemCode;
+import java.io.Serializable;
 import lombok.Data;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -12,7 +13,7 @@ import lombok.Data;
         "mappings"
 })
 @Data
-public class TerminologyMappings {
+public class TerminologyMappings implements Serializable {
     @JsonProperty("openehr")
     private TerminologyMappingsSystemCode openehr;