+++ fhirconnect/src/main/java/com/syntaric/openfhir/fc/schema/terminology/TerminologyMappingsSystemCode.java
@@ -3,6 +3,7 @@ package com.syntaric.openfhir.fc.schema.terminology;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
+import java.io.Serializable;
 import lombok.Data;
 
 @JsonInclude(JsonInclude.Include.NON_NULL)
@@ -12,7 +13,7 @@ import lombok.Data;
         "display"
 })
 @Data
-public class TerminologyMappingsSystemCode {
+public class TerminologyMappingsSystemCode implements Serializable {
     @JsonProperty("system")
     private String system;